Disable until fixed: Bug abuse spells
Those spells are cast by the boss Nefarian, in Blackrock Spire. They force the nearby players to shapeshift into a given form/stance with the Boss as caster, causing them to gain the talent passive auras related to that stance having the Boss as caster.
This cause those passive auras to be not removed when the effect fades, causing them to stack once more every time the spell is cast. A Druid can stack a lot of talent effects increasing his damages to millions.
Couldn't figure any solution by now, but its too problematic to be allowed. Luckly those are the only spells of this kind, and they are not essential so they should be disabled until fixed.